// sanity/types.ts
export type ProjectDetail = {
  _id: string;
  _type: "projectDetail";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  subtitle?: string;
  slug?: Slug;
  description?: string;
  views?: number;
  author?: {
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    [internalGroqTypeReferenceTo]?: "author";
  };
  project?: {
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    [internalGroqTypeReferenceTo]?: "project";
  };
  thumbnail?: string;
  image?: string;
  published?: "pending" | "approved" | "rejected";
  overview?: {
    investor?: string;
    address?: string;
    scale?: string;
    function?: string;
    expense?: string;
    designTeam?: string;
    designYear?: string;
    estimatedTime?: string;
  };
  tags?: string;
  orderIndex?: string;
  pitch?: string;
  onlyShowRouter?: boolean;
  isActived?: boolean;
  isDeleted?: boolean;
  createdAt?: string;
};

export type Project = {
  _id: string;
  _type: "project";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  subtitle?: string;
  slug?: Slug;
  description?: string;
  author?: {
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    [internalGroqTypeReferenceTo]?: "author";
  };
  construction?: Array<{
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    _key: string;
    [internalGroqTypeReferenceTo]?: "construction";
  }>;
  thumbnail?: string;
  image?: string;
  tags?: string;
  orderIndex?: string;
  pitch?: string;
  onlyShowRouter?: boolean;
  isActived?: boolean;
  isDeleted?: boolean;
  createdAt?: string;
};
// ...
export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};

export type Author = {
  _id: string;
  _type: "author";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  type?: "github" | "google";
  id?: string;
  name?: string;
  username?: string;
  email?: string;
  image?: string;
  bio?: string;
  role?: "viewer" | "editor" | "admin";
};

export type Markdown = string;

export type AllSanitySchemaTypes = SanityImagePaletteSwatch | SanityImagePalette | SanityImageDimensions | SanityImageHotspot | SanityImageCrop | SanityFileAsset | SanityImageAsset | SanityImageMetadata | Geopoint | SanityAssetSourceData | ProjectDetail | Project | Design | Construction | Category | Route | Playlist | Startup | Slug | Author | Markdown;
export declare const internalGroqTypeReferenceTo: unique symbol;